New: an ant script to build the war. Had to add some casting in some calls because java has a bug that the eclipse compiler doesn't have and thus the code was compiling with eclipse but not with javac.

+++ src/ch/systemsx/sybit/crkwebui/client/gui/renderers/FinalCallCellRenderer.java
@@ -34,7 +34,7 @@ public class FinalCallCellRenderer implements GridCellRenderer<BaseModel>
 		{
 			String tooltipText = null;
 			
-			int interfaceId = model.get("id");
+			int interfaceId = (Integer)model.get("id");
 			
 			InterfaceItem interfaceItem = mainController.getPdbScoreItem().getInterfaceItem(interfaceId - 1);
